@@ -1,4 +1,6 @@
-﻿using System;
+﻿#nullable disable
+
+using System;
 using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
@@ -61,9 +63,9 @@ namespace Fido2NetLib
             if (Raw.Type != PublicKeyCredentialType.PublicKey)
                 throw new Fido2VerificationException("AssertionResponse Type is not set to public-key");
 
-            if (Raw.Id == null)
+            if (Raw.Id is null)
                 throw new Fido2VerificationException("Id is missing");
-            if (Raw.RawId == null)
+            if (Raw.RawId is null)
                 throw new Fido2VerificationException("RawId is missing");
 
             // 1. If the allowCredentials option was given when this authentication ceremony was initiated, verify that credential.id identifies one of the public key credentials that were listed in allowCredentials.
@@ -99,7 +101,7 @@ namespace Fido2NetLib
 
 
             // 7. Verify that the value of C.type is the string webauthn.get.
-            if (Type != "webauthn.get")
+            if (Type is not "webauthn.get")
                 throw new Fido2VerificationException("AssertionResponse is not type webauthn.get");
 
             // 8. Verify that the value of C.challenge matches the challenge that was sent to the authenticator in the PublicKeyCredentialRequestOptions passed to the get() call.
@@ -144,7 +146,7 @@ namespace Fido2NetLib
 
             // 14. Verify that the values of the client extension outputs in clientExtensionResults and the authenticator extension outputs in the extensions in authData are as expected, considering the client extension input values that were given as the extensions option in the get() call.In particular, any extension identifier values in the clientExtensionResults and the extensions in authData MUST be also be present as extension identifier values in the extensions member of options, i.e., no extensions are present that were not requested. In the general case, the meaning of "are as expected" is specific to the Relying Party and which extensions are in use.
             // todo: Verify this (and implement extensions on options)
-            if (true == authData.HasExtensionsData && ((null == authData.Extensions) || (0 == authData.Extensions.Length))) 
+            if (true == authData.HasExtensionsData && (authData.Extensions is null || (0 == authData.Extensions.Length))) 
                 throw new Fido2VerificationException("Extensions flag present, malformed extensions detected");
             if (false == authData.HasExtensionsData && (null != authData.Extensions)) 
                 throw new Fido2VerificationException("Extensions flag not present, but extensions detected");
@@ -157,7 +159,7 @@ namespace Fido2NetLib
             Buffer.BlockCopy(Raw.Response.AuthenticatorData, 0, data, 0, Raw.Response.AuthenticatorData.Length);
             Buffer.BlockCopy(hashedClientDataJson, 0, data, Raw.Response.AuthenticatorData.Length, hashedClientDataJson.Length);
 
-            if (null == storedPublicKey || 0 == storedPublicKey.Length) 
+            if (storedPublicKey is null || 0 == storedPublicKey.Length) 
                 throw new Fido2VerificationException("Stored public key is null or empty");
             var cpk = new CredentialPublicKey(storedPublicKey);
             if (true != cpk.Verify(data, Signature)) 
