@@ -1,4 +1,6 @@
-﻿using System;
+﻿#nullable disable
+
+using System;
 using System.Linq;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
@@ -26,10 +28,10 @@ namespace Fido2NetLib
 
         public static AuthenticatorAttestationResponse Parse(AuthenticatorAttestationRawResponse rawResponse)
         {
-            if (null == rawResponse || null == rawResponse.Response)
+            if (rawResponse is null || rawResponse.Response is null)
                 throw new Fido2VerificationException("Expected rawResponse, got null");
 
-            if (null == rawResponse.Response.AttestationObject || 0 == rawResponse.Response.AttestationObject.Length)
+            if (rawResponse.Response.AttestationObject is null || 0 == rawResponse.Response.AttestationObject.Length)
                 throw new Fido2VerificationException("Missing AttestationObject");
 
             // 8. Perform CBOR decoding on the attestationObject field of the AuthenticatorAttestationResponse structure to obtain the attestation statement format fmt, the authenticator data authData, and the attestation statement attStmt.
@@ -43,23 +45,23 @@ namespace Fido2NetLib
                 throw new Fido2VerificationException("AttestationObject invalid CBOR", ex);
             }
 
-            if (null == cborAttestation["fmt"]
+            if (cborAttestation["fmt"] is null
                 || CBORType.TextString != cborAttestation["fmt"].Type
-                || null == cborAttestation["attStmt"]
+                || cborAttestation["attStmt"] is null
                 || CBORType.Map != cborAttestation["attStmt"].Type
-                || null == cborAttestation["authData"]
+                || cborAttestation["authData"] is null
                 || CBORType.ByteString != cborAttestation["authData"].Type)
                 throw new Fido2VerificationException("Malformed AttestationObject");
 
             var response = new AuthenticatorAttestationResponse(rawResponse.Response.ClientDataJson)
             {
                 Raw = rawResponse,
-                AttestationObject = new ParsedAttestationObject()
-                {
-                    Fmt = cborAttestation["fmt"].AsString(),
-                    AttStmt = cborAttestation["attStmt"], // convert to dictionary?
-                    AuthData = cborAttestation["authData"].GetByteString()
-                }
+                AttestationObject = new ParsedAttestationObject
+                (
+                    fmt      : cborAttestation["fmt"].AsString(),
+                    attStmt  : cborAttestation["attStmt"], // convert to dictionary?
+                    authData : cborAttestation["authData"].GetByteString()
+                )
             };
             return response;
         }
@@ -73,7 +75,7 @@ namespace Fido2NetLib
             // Above handled in base class constructor
 
             // 3. Verify that the value of C.type is webauthn.create
-            if (Type != "webauthn.create")
+            if (Type is not "webauthn.create")
                 throw new Fido2VerificationException("AttestationResponse is not type webauthn.create");
 
             // 4. Verify that the value of C.challenge matches the challenge that was sent to the authenticator in the create() call.
@@ -82,7 +84,7 @@ namespace Fido2NetLib
             // If Token Binding was used on that TLS connection, also verify that C.tokenBinding.id matches the base64url encoding of the Token Binding ID for the connection.
             BaseVerify(config.Origin, originalOptions.Challenge, requestTokenBindingId);
 
-            if (Raw.Id == null || Raw.Id.Length == 0)
+            if (Raw.Id is null || Raw.Id.Length == 0)
                 throw new Fido2VerificationException("AttestationResponse is missing Id");
 
             if (Raw.Type != PublicKeyCredentialType.PublicKey)
@@ -148,7 +150,7 @@ namespace Fido2NetLib
             var entry = metadataService?.GetEntry(authData.AttestedCredentialData.AaGuid);
 
             // while conformance testing, we must reject any authenticator that we cannot get metadata for
-            if (metadataService?.ConformanceTesting() == true && null == entry && AttestationType.None != attType && "fido-u2f" != AttestationObject.Fmt)
+            if (metadataService?.ConformanceTesting() is true && entry is null && AttestationType.None != attType && AttestationObject.Fmt is not "fido-u2f")
                 throw new Fido2VerificationException("AAGUID not found in MDS test metadata");
 
             if (null != trustPath)
@@ -216,11 +218,18 @@ namespace Fido2NetLib
         /// <summary>
         /// The AttestationObject after CBOR parsing
         /// </summary>
-        public class ParsedAttestationObject
+        public sealed class ParsedAttestationObject
         {
+            public ParsedAttestationObject(string fmt, CBORObject attStmt, byte[] authData)
+            {
+                Fmt = fmt;
+                AttStmt = attStmt;
+                AuthData = authData;
+            }
+
             public string Fmt { get; set; }
-            public byte[] AuthData { get; set; }
             public CBORObject AttStmt { get; set; }
+            public byte[] AuthData { get; set; }
         }
     }
 }
