@@ -12,7 +12,7 @@ namespace Fido2NetLib
         protected readonly string _path;
 
         protected readonly IDictionary<Guid, MetadataBLOBPayloadEntry> _entries;
-        protected MetadataBLOBPayload _blob;
+        protected MetadataBLOBPayload? _blob;
 
         public FileSystemMetadataRepository(string path)
         {
@@ -20,9 +20,9 @@ namespace Fido2NetLib
             _entries = new Dictionary<Guid, MetadataBLOBPayloadEntry>();
         }
 
-        public async Task<MetadataStatement> GetMetadataStatement(MetadataBLOBPayload blob, MetadataBLOBPayloadEntry entry)
+        public async Task<MetadataStatement?> GetMetadataStatement(MetadataBLOBPayload blob, MetadataBLOBPayloadEntry entry)
         {
-            if (_blob == null)
+            if (_blob is null)
                 await GetBLOB();
 
             if (!string.IsNullOrEmpty(entry.AaGuid) && Guid.TryParse(entry.AaGuid, out Guid parsedAaGuid))
