@@ -8,9 +8,9 @@ using PeterO.Cbor;
 
 namespace Fido2NetLib
 {
-    internal class AndroidKey : AttestationVerifier
+    internal sealed class AndroidKey : AttestationVerifier
     {
-        public static byte[] AttestationExtensionBytes(X509ExtensionCollection exts)
+        public static byte[]? AttestationExtensionBytes(X509ExtensionCollection exts)
         {
             foreach (var ext in exts)
             {
@@ -142,14 +142,14 @@ namespace Fido2NetLib
             if (0 == attStmt.Keys.Count || 0 == attStmt.Values.Count)
                 throw new Fido2VerificationException("Attestation format android-key must have attestation statement");
 
-            if (null == Sig || CBORType.ByteString != Sig.Type || 0 == Sig.GetByteString().Length)
+            if ( Sig is null || CBORType.ByteString != Sig.Type || 0 == Sig.GetByteString().Length)
                 throw new Fido2VerificationException("Invalid android-key attestation signature");
             // 2. Verify that sig is a valid signature over the concatenation of authenticatorData and clientDataHash 
             // using the attestation public key in attestnCert with the algorithm specified in alg
-            if (null == X5c || CBORType.Array != X5c.Type || 0 == X5c.Count)
+            if (X5c is null || CBORType.Array != X5c.Type || 0 == X5c.Count)
                 throw new Fido2VerificationException("Malformed x5c in android-key attestation");
 
-            if (null == X5c.Values || 0 == X5c.Values.Count ||
+            if (X5c.Values is null || 0 == X5c.Values.Count ||
                 CBORType.ByteString != X5c.Values.First().Type ||
                 0 == X5c.Values.First().GetByteString().Length)
                 throw new Fido2VerificationException("Malformed x5c in android-key attestation");
@@ -166,7 +166,7 @@ namespace Fido2NetLib
                 throw new Fido2VerificationException("Failed to extract public key from android key: " + ex.Message, ex);
             }
 
-            if (null == Alg || true != Alg.IsNumber)
+            if (Alg is null || true != Alg.IsNumber)
                 throw new Fido2VerificationException("Invalid android key attestation algorithm");
 
             byte[] ecsig;
@@ -188,7 +188,7 @@ namespace Fido2NetLib
 
             // 4. Verify that the attestationChallenge field in the attestation certificate extension data is identical to clientDataHash
             var attExtBytes = AttestationExtensionBytes(androidKeyCert.Extensions);
-            if (null == attExtBytes)
+            if (attExtBytes is null)
                 throw new Fido2VerificationException("Android key attestation certificate contains no AttestationRecord extension");
 
             try
