@@ -9,7 +9,7 @@ namespace Fido2NetLib
         AssertionOptions GetAssertionOptions(
             IEnumerable<PublicKeyCredentialDescriptor> allowedCredentials, 
             UserVerificationRequirement? userVerification, 
-            AuthenticationExtensionsClientInputs extensions = null);
+            AuthenticationExtensionsClientInputs? extensions = null);
 
         Task<AssertionVerificationResult> MakeAssertionAsync(
             AuthenticatorAssertionRawResponse assertionResponse,
@@ -17,24 +17,24 @@ namespace Fido2NetLib
             byte[] storedPublicKey,
             uint storedSignatureCounter,
             IsUserHandleOwnerOfCredentialIdAsync isUserHandleOwnerOfCredentialIdCallback,
-            byte[] requestTokenBindingId = null);
+            byte[]? requestTokenBindingId = null);
 
         Task<Fido2.CredentialMakeResult> MakeNewCredentialAsync(
             AuthenticatorAttestationRawResponse attestationResponse,
             CredentialCreateOptions origChallenge,
             IsCredentialIdUniqueToUserAsyncDelegate isCredentialIdUniqueToUser,
-            byte[] requestTokenBindingId = null);
+            byte[]? requestTokenBindingId = null);
 
         CredentialCreateOptions RequestNewCredential(
             Fido2User user,
             List<PublicKeyCredentialDescriptor> excludeCredentials,
-            AuthenticationExtensionsClientInputs extensions = null);
+            AuthenticationExtensionsClientInputs? extensions = null);
 
         CredentialCreateOptions RequestNewCredential(
             Fido2User user,
             List<PublicKeyCredentialDescriptor> excludeCredentials,
             AuthenticatorSelection authenticatorSelection,
             AttestationConveyancePreference attestationPreference,
-            AuthenticationExtensionsClientInputs extensions = null);
+            AuthenticationExtensionsClientInputs? extensions = null);
     }
 }
