@@ -1,4 +1,6 @@
-﻿using System;
+﻿#nullable disable
+
+using System;
 using System.Reflection;
 using Newtonsoft.Json;
 
